refactor(models): isolate GPT-OSS attention
