+++ src/prime_rl/trainer/models/layers/attn.py
@@ -216,6 +216,6 @@ def substitute_ring_attn(
 
     Qwen3_5GatedFlashAttention._compute_attention = _ring_compute_attention
 
-    from prime_rl.trainer.models.gpt_oss.modeling_gpt_oss import substitute_gpt_oss_ring_attention
+    from prime_rl.trainer.models.gpt_oss.attention import substitute_gpt_oss_ring_attention
 
     substitute_gpt_oss_ring_attention(process_group, heads_k_stride)
