+++ src/prime_rl/trainer/models/layers/ulysses_attn.py
@@ -236,9 +236,9 @@ def substitute_ulysses_attn(
 
     Qwen3_5GatedFlashAttention._compute_attention = _ulysses_compute_attention
 
-    from prime_rl.trainer.models.gpt_oss.modeling_gpt_oss import substitute_gpt_oss_ulysses_attention
+    from prime_rl.trainer.models.gpt_oss.attention import substitute_gpt_oss_ulysses_attention
 
-    substitute_gpt_oss_ulysses_attention(process_group, flash_fn)
+    substitute_gpt_oss_ulysses_attention(process_group)
 
 
 def substitute_hf_ulysses_attn(process_group: dist.ProcessGroup) -> None:
